move highlight functions into shared class
